requirements.txt
google-adk>=0.1.0
google-genai>=0.2.0
tavily-python>=0.3.0
firebase-admin>=6.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0